+++ gammapy/cube/__init__.py
@@ -1,6 +1,3 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Sky cubes (3-dimensional: energy, lon, lat)."""
-from .background import *
-from .exposure import *
-from .make import *
 from .simulate import *
