+++ gammapy/datasets/tests/test_map.py
@@ -5,13 +5,12 @@ from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.coordinates import SkyCoord
 from regions import CircleSkyRegion
-from gammapy.cube import make_map_exposure_true_energy
-from gammapy.datasets import MapDataset, MapDatasetOnOff
+from gammapy.datasets import MapDataset, MapDatasetOnOff, Datasets
 from gammapy.data import GTI
 from gammapy.irf import EffectiveAreaTable2D, EnergyDependentMultiGaussPSF, EDispMap, PSFMap
 from gammapy.maps import Map, MapAxis, WcsGeom, WcsNDMap
 from gammapy.modeling import Fit
-from gammapy.datasets import Datasets
+from gammapy.makers.utils import make_map_exposure_true_energy
 from gammapy.modeling.models import (
     BackgroundModel,
     GaussianSpatialModel,
