+++ gammapy/makers/__init__.py
@@ -1,3 +1,4 @@
 from .background import *
 from .safe import *
 from .spectrum import *
+from .map import *
